Keithley sweep changes
Set maximum safe sweep step size and rate when initializing the Keithley object. All sweeps will be done at the max by default.
